backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miro-backend"
version = "0.1.0"
description = "Backend scaffold for the Miro negotiation coaching platform."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.15.0,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "PyJWT[crypto]>=2.10.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "sqlalchemy>=2.0.39,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.11.0,<1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
